refactor: move TokenScanner, Pipeline and SourceFinder to Utils namespace
These are generic utilities used by both pipelines, not tied to a
specific layer. Moves them to OpenApi\Utils\ and leaves BC subclasses
at the old locations. Test classes moved to matching Utils namespace.
